Rename the simple example to "legacy". The new style is preferred
