+++ src/test/java/examples/simple/legacy/SimpleTableAnnotatedMapper.java
@@ -13,9 +13,9 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package examples.simple;
+package examples.simple.legacy;
 
-import static examples.simple.SimpleTableDynamicSqlSupport.*;
+import static examples.simple.legacy.SimpleTableDynamicSqlSupport.*;
 import static org.mybatis.dynamic.sql.SqlBuilder.*;
 
 import java.util.List;
