+++ src/test/java/examples/simple/newstyle/SimpleTableMapperNewStyleTest.java
@@ -15,7 +15,7 @@
  */
 package examples.simple.newstyle;
 
-import static examples.simple.SimpleTableDynamicSqlSupport.*;
+import static examples.simple.legacy.SimpleTableDynamicSqlSupport.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mybatis.dynamic.sql.SqlBuilder.*;
@@ -43,8 +43,8 @@ import org.mybatis.dynamic.sql.util.mybatis3.MyBatis3CountHelper;
 import org.mybatis.dynamic.sql.util.mybatis3.MyBatis3DeleteHelper;
 import org.mybatis.dynamic.sql.util.mybatis3.MyBatis3SelectListHelper;
 
-import examples.simple.LastName;
-import examples.simple.SimpleTableRecord;
+import examples.simple.legacy.LastName;
+import examples.simple.legacy.SimpleTableRecord;
 
 public class SimpleTableMapperNewStyleTest {
